Correction in cache check (#109)
Incorrect cache usage. According to this logic, if there is no data in the cache for the current URL, this is the normal state. But the plugin will still make the request again.

According to the documentation, you need to check the variable for "false" or use $found in the function.

https://developer.wordpress.org/reference/functions/wp_cache_get/

Signed-off-by: leonidukg <67500820+leonidukg@users.noreply.github.com>